requirements.txt
numpy==1.19.5
Pillow==8.3.1
pkg-resources==0.0.0
scipy==1.5.4
